методы рендоринга вынесены в render.js
